ENGR00133669: MXC: Ensure some clock APIs can be called in process context only.
The added checks will ensure that clk_enable() and clk_disable() are called
only in a process context(). The main reason behind doing this is to allow
for voltage scaling as part of clk_enable()/clk_disable() functions.
Voltage scaling uses the Regulator APIs which cannot be called in an
ISR context.
This also reduces the amount of work done in an ISR context.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>