ENGR00139054-2: MX50- Simplify DDR frequency change code.
The DDR init code is changing frequently and is modified based on the
memory vendor. Change the DDR frequency code in the BSP so that is
reads the DDR settings initialized by the bootlader. Use these settings
when switching between 24Mhz and 200/266MHz.
Also simplify the bus frequency driver, move MX50 DDR freq change to a
different file.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>