ENGR00141092 usb: fix some tiny bugs for usb
For android system, the gadget driver is built in, when the usb
device is on otg port, the udc suspended counter will be error at
the situation. This patch fix it, if this situation is met, the
udc suspended counter will be not added.

Using A VBus Valid to judge whether the usb cable is connected, as
some boards' vbus may higher than 0.8v(B Session Valid)

Signed-off-by: Peter Chen <peter.chen@freescale.com>